requirements.txt
meshtastic
pypubsub
